+++ lib/buyer_app/buyer_order_details/view/widgets/buyer_order_products_list_item.dart
@@ -143,7 +143,7 @@ class BuyerOrderProductsListItem extends StatelessWidget {
                         OrderQuantityChangeButton(text: "-", onTap: onRemove),
                         const SizedBox(width: 8),
                         OrderQuantityDisplayTextView(
-                          value: item.cartQty ?? 0,
+                          value: (item.cartQty ?? 0).toInt(),
                           width: 52,
                           height: 30,
                         ),
