+++ lib/pages/user_orders/basket/view/widgets/basket_items_list.dart
@@ -31,7 +31,8 @@ class _BasketItemsListState extends State<BasketItemsList> {
               final product = controller.cartList[index];
               final pageController = PageController();
 
-              final outOfStockCount = product.cartQty ?? 0 - (product.qty ?? 0);
+              final outOfStockCount = (product.cartQty ?? 0.0) - (product.qty ?? 0.0);
+
               return InkWell(
                 onTap: (){
                   var arguments = {
@@ -173,7 +174,7 @@ class _BasketItemsListState extends State<BasketItemsList> {
                                       }),
                                       SizedBox(width: 8),
                                       OrderQuantityDisplayTextView(
-                                        value: product.cartQty ?? 0,
+                                        value: (product.cartQty ?? 0).toInt(),
                                         width: 52,
                                         height: 30,
                                       ),
@@ -203,7 +204,7 @@ class _BasketItemsListState extends State<BasketItemsList> {
                         // OUT OF STOCK MESSAGE
                         if ((product.qty ?? 0) >= 0)
                           OutOfStockBanner(
-                            itemCount: outOfStockCount,
+                            itemCount: (outOfStockCount).toInt(),
                             deliveryDays: 5,
                           ),
                       ],
